+++ client/components/BubbleChart.jsx
@@ -1,17 +1,43 @@
 import React from 'react';
-import { Bubble } from 'react-chartjs';
+import { Bubble } from 'react-chartjs-2';
 
-class BubbleChart extends React.Component {
-    constructor(data, options) {
-        super()
-    }
+const data = {
+    labels: ['January'],
+    datasets: [
+        {
+            label: 'My First dataset',
+            fill: false,
+            lineTension: 0.1,
+            backgroundColor: 'rgba(75,192,192,0.4)',
+            borderColor: 'rgba(75,192,192,1)',
+            borderCapStyle: 'butt',
+            borderDash: [],
+            borderDashOffset: 0.0,
+            borderJoinStyle: 'miter',
+            pointBorderColor: 'rgba(75,192,192,1)',
+            pointBackgroundColor: '#fff',
+            pointBorderWidth: 1,
+            pointHoverRadius: 5,
+            pointHoverBackgroundColor: 'rgba(75,192,192,1)',
+            pointHoverBorderColor: 'rgba(220,220,220,1)',
+            pointHoverBorderWidth: 2,
+            pointRadius: 1,
+            pointHitRadius: 10,
+            data: [{ x: 10, y: 20, r: 5 }]
+        }
+    ]
+};
+
+export default React.createClass({
+    displayName: 'BubbleChart',
+
     render() {
         return (
             <div>
-                <Bubble data={this.props.data} options={this.props.options} width="600" height="250"/>
+                <h2>Speed Progress</h2>
+                <p>Tracks your past runs by date and predicts future progress</p>
+                <Bubble data={data} />
             </div>
-        )
-    }   
-}
-
-export default BubbleChart
+        );
+    }
+});
