+++ client/components/LineChart.jsx
@@ -1,17 +1,64 @@
 import React from 'react';
-import { Line } from 'react-chartjs';
+import { Line } from 'react-chartjs-2';
+import regression from 'regression';
 
-class LineChart extends React.Component {
-    constructor(data, options) {
-        super()
-    }
+var input = [[1,4],[1.5,null],[2,26],[0.5,2],[2.5, null],[3, null]];
+
+var sorted = input.sort(function(a,b) {
+  return a[0] - b[0];
+});
+
+var result = regression('linear', sorted);
+
+var unzip = function(array, index) {
+  var output = [];
+  for (var i = 0; i < array.length; i++) {
+    output.push(array[i][index]);
+  }
+  return output;
+}
+
+var xArray = unzip(result.points, 0);
+var yArray = unzip(result.points, 1);
+
+var data = {
+    labels: xArray,
+    datasets: [
+        {
+            label: "My First dataset",
+            fill: false,
+            lineTension: 0.1,
+            backgroundColor: "rgba(75,192,192,0.4)",
+            borderColor: "rgba(75,192,192,1)",
+            borderCapStyle: 'butt',
+            borderDash: [],
+            borderDashOffset: 0.0,
+            borderJoinStyle: 'miter',
+            pointBorderColor: "rgba(75,192,192,1)",
+            pointBackgroundColor: "#fff",
+            pointBorderWidth: 1,
+            pointHoverRadius: 5,
+            pointHoverBackgroundColor: "rgba(75,192,192,1)",
+            pointHoverBorderColor: "rgba(220,220,220,1)",
+            pointHoverBorderWidth: 2,
+            pointRadius: 1,
+            pointHitRadius: 10,
+            data: yArray,
+            spanGaps: false,
+        }
+    ]
+};
+
+export default React.createClass({
+    displayName: 'LineChart',
+
     render() {
         return (
             <div>
-                <Line data={this.props.data} options={this.props.options} width="600" height="250"/>
+                <h2>Time Estimates</h2>
+                <p>Predicts the time it will take you to run a number of miles based on your history</p>
+                <Line data={data} />
             </div>
-        )
+        );
     }
-}
-
-export default LineChart
+});
