+++ package.json
@@ -20,11 +20,11 @@
   "homepage": "https://github.com/BrandonKleiman/Thesis#readme",
   "dependencies": {
     "auth0-lock": "^10.14.0",
-    "axios": "^0.15.3",
+    "axios": "^0.16.0",
     "body-parser": "^1.17.1",
-    "chart.js": "^1.1.1",
+    "chart.js": "^2.5.0",
     "classnames": "^2.2.5",
-    "cors": "^2.8.1",
+    "cors": "^2.8.3",
     "dotenv": "^4.0.0",
     "express": "^4.15.2",
     "history": "^4.6.1",
@@ -34,12 +34,12 @@
     "mysql": "^2.13.0",
     "path": "^0.12.7",
     "react": "^15.4.2",
-    "react-chartjs": "^0.8.0",
+    "react-chartjs-2": "^2.0.5",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.3",
     "react-router": "^4.0.0",
     "redux": "^3.6.0",
-    "redux-axios-middleware": "^3.1.2",
+    "redux-axios-middleware": "^4.0.0",
     "redux-thunk": "^2.2.0",
     "regression": "^1.4.0",
     "request": "^2.81.0",
@@ -48,21 +48,21 @@
     "sequelize": "^3.30.4"
   },
   "devDependencies": {
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.4.0",
+    "babel-core": "^6.24.0",
+    "babel-loader": "^6.4.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-es2015": "^6.24.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.22.0",
     "babel-preset-stage-2": "^6.22.0",
-    "eslint": "^3.18.0",
+    "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.10.3",
     "nodemon": "^1.11.0",
-    "webpack": "^2.2.1"
+    "webpack": "^2.3.3"
   }
 }
